pset8 Readme
Hi! This is the webpage that I created for my submission to problem set 8 in CS50 (HTML, CSS and Javascript). This website will allow you to navigate through the different exercises recommended for different body parts. The webpage will welcome the user to the webpage wih an alert, and will then greet the user by their actual name when they submit it on the homepage.