pset9 Sell.html
This is the html for the "sell" directory of the website. This page will allow the user to sell stocks which they already have. The logic for this is carried out in the "sell" secgtion of application.py. It will also check if the user has enough of the selected stock available to sell.